+++ src/django_bird/plugins/manager.py
@@ -13,6 +13,7 @@ pm.load_setuptools_entrypoints("django_bird")
 
 DEFAULT_PLUGINS: list[str] = [
     "django_bird.staticfiles",
+    "django_bird.templates",
 ]
 
 for plugin in DEFAULT_PLUGINS: